fix: skip pip-audit when VIRTUAL_ENV is set but project has no Python manifest
Closes #207
